fix: ignore clippy false-positive and add tests
