+++ manta-accounting/Cargo.toml
@@ -65,5 +65,5 @@ rayon = { version = "1.5.1", optional = true, default-features = false }
 statrs = { version = "0.15.0", optional = true, default-features = false }
 
 [dev-dependencies]
-rand = "0.8.4"
+manta-crypto = { path = "../manta-crypto", default-features = false, features = ["getrandom"] }
