+++ manta-accounting/src/asset.rs
@@ -1134,12 +1134,12 @@ where
 #[cfg(test)]
 mod test {
     use super::*;
-    use rand::thread_rng;
+    use manta_crypto::rand::OsRng;
 
     /// Tests asset conversion into and from bytes.
     #[test]
     fn asset_into_and_from_bytes() {
-        let mut rng = thread_rng();
+        let mut rng = OsRng;
         let asset = Asset::gen(&mut rng);
         assert_eq!(asset, Asset::from_bytes(asset.into_bytes()));
         let mut asset_bytes = [0; Asset::SIZE];
@@ -1150,7 +1150,7 @@ mod test {
     /// Tests asset arithmetic.
     #[test]
     fn asset_arithmetic() {
-        let mut rng = thread_rng();
+        let mut rng = OsRng;
         let mut asset = Asset::zero(rng.gen());
         let value = rng.gen();
         let _ = asset + value;
